refactor: move to japa plugin
